+++ src/app/app.component.html
@@ -1,7 +1,3 @@
 <app-navbar></app-navbar>
-
-<h1> Où nous trouver ?</h1>
-<div class="container mt-5" id="mapa-mapbox"></div>
-
 <router-outlet></router-outlet>
 <app-footer></app-footer>
