+++ src/app/app.component.scss
@@ -1,16 +1,1 @@
-#mapa-mapbox {
-    width: 50%;
-    height: 50%;
-}
 
-.marker {
-    display: block;
-    border: none;
-    border-radius: 50%;
-    cursor: pointer;
-    padding: 0
-}
-
-h1 {
-    text-align: center;
-}
