+++ src/app/home/home.component.html
@@ -1,1 +1,1 @@
-<p>home works!</p>
+<app-localisation></app-localisation>
